requirements.txt
torch
torchvision
monai
mlflow
boto3
python-dotenv
fastapi
uvicorn
python-multipart
numpy
opencv-python